Feature/profile service (#4)
* [BE] profile service

+++ backend/profile-service/src/main/kotlin/com/pro100byte/util/JwtUtil.kt
@@ -46,6 +46,6 @@ class JwtUtil(
 
     fun decodeId(token: String): Long {
         val body = Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).body
-        return body.get(ID_CLAIM, Long::class.java)
+        return body.get(ID_CLAIM, Integer::class.java).toLong()
     }
 }
